requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
scikit-learn
xgboost
pandas